Get data on a user

Retrieve comprehensive data for a user in the current FPL season, obtained via the user entry history endpoint.

fpl_get_user(user_id)

Arguments

user_id
user ID (found in a user's team page URL on the main FPL site).

Value

a list

Examples

fpl_get_user(user_id = 3808385)

#> $chips
#>   played_time_formatted status     name                 time chip   entry event
#> 1          19 Oct 01:33 played wildcard 2018-10-19T00:33:57Z    1 3808385     9
#> 2          30 Mar 02:02 played  freehit 2019-03-30T02:02:16Z    3 3808385    32
#> 3          12 Apr 17:43 played wildcard 2019-04-12T16:43:53Z    2 3808385    34
#> 4          20 Apr 02:36 played   bboost 2019-04-20T01:36:24Z    5 3808385    35
#> 5          26 Apr 17:35 played      3xc 2019-04-26T16:35:01Z    4 3808385    36
#> 
#> $entry
#> $entry$id
#> [1] 3808385
#> 
#> $entry$player_first_name
#> [1] "Amos"
#> 
#> $entry$player_last_name
#> [1] "Bastian"
#> 
#> $entry$player_region_id
#> [1] 152
#> 
#> $entry$player_region_name
#> [1] "Netherlands"
#> 
#> $entry$player_region_short_iso
#> [1] "NL"
#> 
#> $entry$summary_overall_points
#> [1] 2414
#> 
#> $entry$summary_overall_rank
#> [1] 4552
#> 
#> $entry$summary_event_points
#> [1] 85
#> 
#> $entry$summary_event_rank
#> [1] 129877
#> 
#> $entry$joined_seconds
#> [1] 30947
#> 
#> $entry$current_event
#> [1] 38
#> 
#> $entry$total_transfers
#> [1] 34
#> 
#> $entry$total_loans
#> [1] 0
#> 
#> $entry$total_loans_active
#> [1] 0
#> 
#> $entry$transfers_or_loans
#> [1] "transfers"
#> 
#> $entry$deleted
#> [1] FALSE
#> 
#> $entry$email
#> [1] FALSE
#> 
#> $entry$joined_time
#> [1] "2018-08-09T22:44:21Z"
#> 
#> $entry$name
#> [1] "( ͡° \u035cʖ ͡°)"
#> 
#> $entry$bank
#> [1] 7
#> 
#> $entry$value
#> [1] 1049
#> 
#> $entry$kit
#> [1] "{\"kit_shirt_type\":\"plain\",\"kit_shirt_base\":\"#ff0000\",\"kit_shirt_sleeves\":\"#ff0000\",\"kit_shirt_secondary\":\"#e1e1e1\",\"kit_shirt_logo\":\"none\",\"kit_shorts\":\"#000000\",\"kit_socks_type\":\"plain\",\"kit_socks_base\":\"#ffffff\",\"kit_socks_secondary\":\"#e1e1e1\"}"
#> 
#> $entry$event_transfers
#> [1] 0
#> 
#> $entry$event_transfers_cost
#> [1] 0
#> 
#> $entry$extra_free_transfers
#> [1] 0
#> 
#> $entry$strategy
#> NULL
#> 
#> $entry$favourite_team
#> [1] 14
#> 
#> $entry$started_event
#> [1] 1
#> 
#> $entry$player
#> [1] 7425806
#> 
#> 
#> $leagues
#> $leagues$cup
#> list()
#> 
#> $leagues$h2h
#> list()
#> 
#> $leagues$classic
#>        id entry_rank entry_last_rank entry_movement entry_change
#> 1      14        792            1390             up          598
#> 2     172         24              43             up           19
#> 3     275       4549            7993             up         3444
#> 4     313       4552            7998             up         3446
#> 5    1137        300             488             up          188
#> 6  171842        970            1572             up          602
#> 7  202820          1               1           same           NA
#> 8  489511          6               8             up            2
#> 9  633353          2               3             up            1
#> 10 890172          1               1           same           NA
#>    entry_can_leave entry_can_admin entry_can_invite entry_can_forum entry_code
#> 1            FALSE           FALSE            FALSE           FALSE         NA
#> 2            FALSE           FALSE            FALSE           FALSE         NA
#> 3            FALSE           FALSE            FALSE           FALSE         NA
#> 4            FALSE           FALSE            FALSE           FALSE         NA
#> 5             TRUE           FALSE            FALSE            TRUE         NA
#> 6             TRUE           FALSE            FALSE            TRUE         NA
#> 7             TRUE           FALSE            FALSE            TRUE         NA
#> 8             TRUE           FALSE            FALSE            TRUE         NA
#> 9             TRUE           FALSE            FALSE            TRUE         NA
#> 10            TRUE           FALSE            FALSE            TRUE         NA
#>                                    name short_name              created closed
#> 1                               Man Utd    team-14 2018-07-05T12:12:23Z  FALSE
#> 2                           Netherlands region-152 2018-07-05T12:12:23Z  FALSE
#> 3                            Gameweek 1    event-1 2018-07-05T12:12:24Z  FALSE
#> 4                               Overall    overall 2018-07-05T12:12:24Z  FALSE
#> 5  Official /r/FantasyPL Classic League       <NA> 2018-07-05T15:01:19Z  FALSE
#> 6              The League of Doing Bits       <NA> 2018-07-19T20:53:13Z  FALSE
#> 7                         Elites League       <NA> 2018-07-21T17:00:43Z  FALSE
#> 8                          /r/reddevils       <NA> 2018-08-03T11:00:36Z  FALSE
#> 9                  Steem Fantasy League       <NA> 2018-08-07T00:03:18Z  FALSE
#> 10                          AJ's Angels       <NA> 2018-08-10T08:15:37Z  FALSE
#>    forum_disabled make_code_public rank size league_type _scoring
#> 1           FALSE            FALSE   NA   NA           s        c
#> 2           FALSE            FALSE   NA   NA           s        c
#> 3           FALSE            FALSE   NA   NA           s        c
#> 4           FALSE            FALSE   NA   NA           s        c
#> 5           FALSE            FALSE   NA   NA           x        c
#> 6           FALSE            FALSE   NA   NA           x        c
#> 7           FALSE            FALSE   NA   NA           x        c
#> 8           FALSE            FALSE   NA   NA           x        c
#> 9           FALSE            FALSE   NA   NA           x        c
#> 10          FALSE            FALSE   NA   NA           x        c
#>    reprocess_standings admin_entry start_event
#> 1                FALSE          NA           1
#> 2                FALSE          NA           1
#> 3                FALSE          NA           1
#> 4                FALSE          NA           1
#> 5                FALSE        3027           1
#> 6                FALSE      769483           1
#> 7                FALSE      903524           1
#> 8                FALSE     1884352           1
#> 9                FALSE     2779525           1
#> 10               FALSE        9346           1
#> 
#> 
#> $season
#>         id season_name total_points    rank season  player
#> 1  7408636     2010/11         1726  923120      5 7425806
#> 2 14217000     2011/12         1706 1219063      6 7425806
#> 3 19982494     2012/13         1745 1243765      7 7425806
#> 4 21729639     2013/14         2019  965336      8 7425806
#> 5 32730885     2014/15         2016  174160      9 7425806
#> 6 36051053     2015/16         2212   32360     10 7425806
#> 7 40342198     2016/17         2146  140367     11 7425806
#> 8 46149067     2017/18         2272   29397     12 7425806
#> 
#> $history
#>           id movement points total_points    rank rank_sort overall_rank
#> 1    3808385      new     89           89   83755     93903        83755
#> 2    8364697       up     88          177  312528    323404        59210
#> 3   13473911     down     42          219 4060967   4085020       125227
#> 4   18847781       up     64          283  251716    262169        45346
#> 5   24352721       up     75          358  306632    312891        20025
#> 6   29988731     down     51          409 3068229   3105878        37672
#> 7   35624741     down     47          456 3473846   3505687        78286
#> 8   41326286     down     48          504 3026744   3045561       146361
#> 9   47093366     down     36          540 4210058   4265087       187042
#> 10  52860446       up     89          629  188422    195824        80466
#> 11  58693061       up     85          714  105391    108833        34776
#> 12  64525676     down     47          761 3354566   3386857        52764
#> 13  70423826       up     67          828  575810    584503        24273
#> 14  76321976       up     63          891  795208    806635        18369
#> 15  82285661       up     62          953  678407    690475        14287
#> 16  88249346     down     37          986 5102264   5111660        46158
#> 17  94213031       up     65         1051  206786    218631        28773
#> 18 100176716     down     42         1093 4622608   4665567        47581
#> 19 106205936     down     55         1148 3670111   3729603        83664
#> 20 112235156       up     60         1208 1490998   1504404        69620
#> 21 118329911       up     62         1270 1442968   1512296        59751
#> 22 124424666       up     73         1343  143063    148442        42433
#> 23 130584956     down     65         1404 1480275   1488570        49225
#> 24 136745246       up     46         1450 2256239   2329549        47635
#> 25 142905536       up     96         1546  379344    383921        26180
#> 26 149131361       up     75         1621  839850    854647        22424
#> 27 155357186     down     35         1656 3255445   3276968        26444
#> 28 161583011       up     63         1719 1564502   1615461        21931
#> 29 167808836       up     47         1766 1693662   1712685        20278
#> 30 174100196     down     48         1814 3092516   3110261        24937
#> 31 180391556     down     32         1846 1978438   1986990        27850
#> 32 186682916     down     73         1919 2923238   2986544        41371
#> 33 192974276       up     79         1998   92701     95232        26518
#> 34 199265636       up     55         2053 2849318   2915682        24844
#> 35 205557190       up    115         2168   43438     44746        14152
#> 36 211914085       up    110         2278  124935    129383         8972
#> 37 218270980       up     51         2329 2622716   2635215         7998
#> 38 224627875       up     85         2414  129877    132717         4552
#>    targets event_transfers event_transfers_cost value points_on_bench bank
#> 1       NA               0                    0  1000              15    5
#> 2       NA               0                    0  1005               4    5
#> 3       NA               2                    0  1012               6    5
#> 4       NA               1                    0  1020               0    6
#> 5       NA               1                    0  1022               0    4
#> 6       NA               1                    0  1022               0   12
#> 7       NA               1                    0  1026               1   13
#> 8       NA               1                    0  1025               5    0
#> 9       NA               0                    0  1028               1   15
#> 10      NA               0                    0  1032               1   15
#> 11      NA               1                    0  1033              14   18
#> 12      NA               1                    0  1037              10   16
#> 13      NA               1                    0  1041               9   17
#> 14      NA               2                    0  1041               8    1
#> 15      NA               1                    0  1044               0    2
#> 16      NA               2                    4  1047               1    1
#> 17      NA               0                    0  1049               8    1
#> 18      NA               0                    0  1053               1    1
#> 19      NA               0                    0  1050               8    1
#> 20      NA               2                    0  1049               4   26
#> 21      NA               0                    0  1052               8   26
#> 22      NA               1                    0  1051               3    5
#> 23      NA               3                    4  1056               1   22
#> 24      NA               0                    0  1059               2   22
#> 25      NA               1                    0  1062               1   22
#> 26      NA               1                    0  1063               5   43
#> 27      NA               2                    0  1067               0   13
#> 28      NA               0                    0  1068               2   13
#> 29      NA               2                    0  1067               6    1
#> 30      NA               1                    0  1068              17   11
#> 31      NA               1                    0  1065               0   16
#> 32      NA               0                    0  1064               7   16
#> 33      NA               1                    0  1059               2   42
#> 34      NA               0                    0  1047              13   18
#> 35      NA               1                    0  1049               0    7
#> 36      NA               0                    0  1053               7    7
#> 37      NA               2                    0  1055              12    0
#> 38      NA               1                    0  1056               5    7
#>      entry event
#> 1  3808385     1
#> 2  3808385     2
#> 3  3808385     3
#> 4  3808385     4
#> 5  3808385     5
#> 6  3808385     6
#> 7  3808385     7
#> 8  3808385     8
#> 9  3808385     9
#> 10 3808385    10
#> 11 3808385    11
#> 12 3808385    12
#> 13 3808385    13
#> 14 3808385    14
#> 15 3808385    15
#> 16 3808385    16
#> 17 3808385    17
#> 18 3808385    18
#> 19 3808385    19
#> 20 3808385    20
#> 21 3808385    21
#> 22 3808385    22
#> 23 3808385    23
#> 24 3808385    24
#> 25 3808385    25
#> 26 3808385    26
#> 27 3808385    27
#> 28 3808385    28
#> 29 3808385    29
#> 30 3808385    30
#> 31 3808385    31
#> 32 3808385    32
#> 33 3808385    33
#> 34 3808385    34
#> 35 3808385    35
#> 36 3808385    36
#> 37 3808385    37
#> 38 3808385    38
#>