Get entry data on a user

Retrieve entry information for a user in the current FPL season, obtained via the user entry history endpoint.

fpl_get_user_entry(user_id)

Arguments

user_id
user ID (found in a user's team page URL on the main FPL site).

Value

a list

Examples

fpl_get_user_entry(user_id = 3808385)

#> $id
#> [1] 3808385
#> 
#> $player_first_name
#> [1] "Amos"
#> 
#> $player_last_name
#> [1] "Bastian"
#> 
#> $player_region_id
#> [1] 152
#> 
#> $player_region_name
#> [1] "Netherlands"
#> 
#> $player_region_short_iso
#> [1] "NL"
#> 
#> $summary_overall_points
#> [1] 2414
#> 
#> $summary_overall_rank
#> [1] 4552
#> 
#> $summary_event_points
#> [1] 85
#> 
#> $summary_event_rank
#> [1] 129877
#> 
#> $joined_seconds
#> [1] 77295
#> 
#> $current_event
#> [1] 38
#> 
#> $total_transfers
#> [1] 34
#> 
#> $total_loans
#> [1] 0
#> 
#> $total_loans_active
#> [1] 0
#> 
#> $transfers_or_loans
#> [1] "transfers"
#> 
#> $deleted
#> [1] FALSE
#> 
#> $email
#> [1] FALSE
#> 
#> $joined_time
#> [1] "2018-08-09T22:44:21Z"
#> 
#> $name
#> [1] "( ͡° \u035cʖ ͡°)"
#> 
#> $bank
#> [1] 7
#> 
#> $value
#> [1] 1049
#> 
#> $kit
#> [1] "{\"kit_shirt_type\":\"plain\",\"kit_shirt_base\":\"#ff0000\",\"kit_shirt_sleeves\":\"#ff0000\",\"kit_shirt_secondary\":\"#e1e1e1\",\"kit_shirt_logo\":\"none\",\"kit_shorts\":\"#000000\",\"kit_socks_type\":\"plain\",\"kit_socks_base\":\"#ffffff\",\"kit_socks_secondary\":\"#e1e1e1\"}"
#> 
#> $event_transfers
#> [1] 0
#> 
#> $event_transfers_cost
#> [1] 0
#> 
#> $extra_free_transfers
#> [1] 0
#> 
#> $strategy
#> NULL
#> 
#> $favourite_team
#> [1] 14
#> 
#> $started_event
#> [1] 1
#> 
#> $player
#> [1] 7425806
#>