Detailed data for a player in the current FPL season

Returns a tibble containing gameweek-level data for a given player in the current FPL season.

playerDetailed(player_id)

Arguments

player_id
id field from players tibble for a desired player.

Examples

playerDetailed(player_id = 1)

#> # A tibble: 38 x 54
#>       id         kickoff_time kickoff_time_formatted team_h_score team_a_score
#>    <int>                <chr>                  <chr>        <int>        <int>
#>  1     1 2016-08-14T15:00:00Z           14 Aug 16:00            3            4
#>  2   526 2016-08-20T16:30:00Z           20 Aug 17:30            0            0
#>  3  1066 2016-08-27T14:00:00Z           27 Aug 15:00            1            3
#>  4  1616 2016-09-10T14:00:00Z           10 Sep 15:00            2            1
#>  5  2197 2016-09-17T14:00:00Z           17 Sep 15:00            1            4
#>  6  2781 2016-09-24T16:30:00Z           24 Sep 17:30            3            0
#>  7  3367 2016-10-02T15:30:00Z           02 Oct 16:30            0            1
#>  8  3956 2016-10-15T14:00:00Z           15 Oct 15:00            3            2
#>  9  4548 2016-10-22T14:00:00Z           22 Oct 15:00            0            0
#> 10  5142 2016-10-29T11:30:00Z           29 Oct 12:30            1            4
#> # ... with 28 more rows, and 49 more variables: was_home <lgl>, round <int>,
#> #   total_points <int>, transfers_balance <int>, selected <int>,
#> #   transfers_in <int>, transfers_out <int>, loaned_in <int>, loaned_out <int>,
#> #   minutes <int>, goals_scored <int>, assists <int>, clean_sheets <int>,
#> #   goals_conceded <int>, own_goals <int>, penalties_saved <int>,
#> #   penalties_missed <int>, yellow_cards <int>, red_cards <int>, saves <int>,
#> #   bonus <int>, bps <int>, influence <dbl>, creativity <dbl>, threat <dbl>,
#> #   ict_index <dbl>, ea_index <int>, open_play_crosses <int>,
#> #   big_chances_created <int>, clearances_blocks_interceptions <int>,
#> #   recoveries <int>, key_passes <int>, tackles <int>, winning_goals <int>,
#> #   attempted_passes <int>, completed_passes <int>, penalties_conceded <int>,
#> #   big_chances_missed <int>, errors_leading_to_goal <int>,
#> #   errors_leading_to_goal_attempt <int>, tackled <int>, offside <int>,
#> #   target_missed <int>, fouls <int>, dribbles <int>, fixture <int>,
#> #   opponent_team <chr>, price <dbl>, player_id <int>

playerDetailed(player_id = 54)

#> # A tibble: 38 x 54
#>       id         kickoff_time kickoff_time_formatted team_h_score team_a_score
#>    <int>                <chr>                  <chr>        <int>        <int>
#>  1    54 2016-08-13T14:00:00Z           13 Aug 15:00            0            1
#>  2   579 2016-08-20T14:00:00Z           20 Aug 15:00            2            0
#>  3  1119 2016-08-27T14:00:00Z           27 Aug 15:00            3            0
#>  4  1669 2016-09-10T14:00:00Z           10 Sep 15:00            1            1
#>  5  2250 2016-09-17T14:00:00Z           17 Sep 15:00            3            0
#>  6  2834 2016-09-26T19:00:00Z           26 Sep 20:00            2            0
#>  7  3420 2016-10-02T15:30:00Z           02 Oct 16:30            0            1
#>  8  4009 2016-10-16T15:00:00Z           16 Oct 16:00            3            1
#>  9  4601 2016-10-22T14:00:00Z           22 Oct 15:00            2            1
#> 10  5195 2016-10-29T14:00:00Z           29 Oct 15:00            0            0
#> # ... with 28 more rows, and 49 more variables: was_home <lgl>, round <int>,
#> #   total_points <int>, transfers_balance <int>, selected <int>,
#> #   transfers_in <int>, transfers_out <int>, loaned_in <int>, loaned_out <int>,
#> #   minutes <int>, goals_scored <int>, assists <int>, clean_sheets <int>,
#> #   goals_conceded <int>, own_goals <int>, penalties_saved <int>,
#> #   penalties_missed <int>, yellow_cards <int>, red_cards <int>, saves <int>,
#> #   bonus <int>, bps <int>, influence <dbl>, creativity <dbl>, threat <dbl>,
#> #   ict_index <dbl>, ea_index <int>, open_play_crosses <int>,
#> #   big_chances_created <int>, clearances_blocks_interceptions <int>,
#> #   recoveries <int>, key_passes <int>, tackles <int>, winning_goals <int>,
#> #   attempted_passes <int>, completed_passes <int>, penalties_conceded <int>,
#> #   big_chances_missed <int>, errors_leading_to_goal <int>,
#> #   errors_leading_to_goal_attempt <int>, tackled <int>, offside <int>,
#> #   target_missed <int>, fouls <int>, dribbles <int>, fixture <int>,
#> #   opponent_team <chr>, price <dbl>, player_id <int>