Historic FPL seasons overview for a player in the current season

Returns a tibble containing a season history overview for a given player in the current FPL season.

playerSeasons(player_id)

Arguments

player_id
id field from players tibble for a desired player.

Examples

playerSeasons(player_id = 1)

#> # A tibble: 3 x 26
#>      id season_name element_code total_points minutes goals_scored assists
#>   <int>       <chr>        <int>        <int>   <int>        <int>   <int>
#> 1  5424     2014/15        48844           76    1620            0       0
#> 2  6096     2015/16        48844           20     360            0       0
#> 3  6818     2016/17        48844            2     143            0       0
#> # ... with 19 more variables: clean_sheets <int>, goals_conceded <int>,
#> #   own_goals <int>, penalties_saved <int>, penalties_missed <int>,
#> #   yellow_cards <int>, red_cards <int>, saves <int>, bonus <int>, bps <int>,
#> #   influence <dbl>, creativity <dbl>, threat <dbl>, ict_index <dbl>,
#> #   ea_index <int>, season <int>, start_price <dbl>, end_price <dbl>,
#> #   player_id <dbl>

playerSeasons(player_id = 54)

#> Warning: Unknown or uninitialised column: 'influence'.

#> Warning: Unknown or uninitialised column: 'creativity'.

#> Warning: Unknown or uninitialised column: 'threat'.

#> Warning: Unknown or uninitialised column: 'ict_index'.

#> Warning: Unknown or uninitialised column: 'start_cost'.

#> Warning: Unknown or uninitialised column: 'end_cost'.

#> Error in `$<-.data.frame`(`*tmp*`, "player_id", value = 54): replacement has 1 row, data has 0