Detailed data for a user's player picks for a gameweek in the current FPL season

Returns a tibble containing player data for a given user's selection for a gameweek during the current FPL season.

userPicks(user_id, gameweek)

Arguments

user_id
FPL user ID (retrieve from the corresponding user's https://fantasy.premierleague.com team page URL).

gameweek
FPL competition gameweek (1-38).

Examples

userPicks(user_id = 123, gameweek = 10)

#> # A tibble: 15 x 57
#>    round fixture element position player_name   player_team
#>  * <int>   <int>   <int>    <int>       <chr>         <chr>
#>  1    10      92     450        1      Adrián       Everton
#>  2    10      97       3        2   Koscielny    Sunderland
#>  3    10      94     278        3      Gibson   Bournemouth
#>  4    10     100     436        4     McAuley      Man City
#>  5    10      91     104        5        Dann     Liverpool
#>  6    10      97      12        6     Sánchez    Sunderland
#>  7    10      96     333        7     Shaqiri       Swansea
#>  8    10      98     402        8         Son     Leicester
#>  9    10      95     314        9      Austin       Chelsea
#> 10    10      95      97       10       Costa   Southampton
#> 11    10      92     143       11      Lukaku      West Ham
#> 12    10     100     430       12      Foster      Man City
#> 13    10      97      13       13     Walcott    Sunderland
#> 14    10      94      42       14   Stanislas Middlesbrough
#> 15    10      92     495       15     Holgate      West Ham
#> # ... with 51 more variables: player_team_strength <int>, opponent_team <chr>,
#> #   opponent_team_strength <int>, kickoff_time <chr>, team_h_score <int>,
#> #   team_a_score <int>, was_home <lgl>, is_captain <lgl>,
#> #   is_vice_captain <lgl>, multiplier <int>, total_points <int>, price <dbl>,
#> #   selected <int>, transfers_in <int>, transfers_out <int>, minutes <int>,
#> #   goals_scored <int>, assists <int>, clean_sheets <int>,
#> #   goals_conceded <int>, own_goals <int>, penalties_saved <int>,
#> #   penalties_missed <int>, yellow_cards <int>, red_cards <int>, saves <int>,
#> #   bonus <int>, bps <int>, influence <dbl>, creativity <dbl>, threat <dbl>,
#> #   ict_index <dbl>, ea_index <int>, open_play_crosses <int>,
#> #   big_chances_created <int>, clearances_blocks_interceptions <int>,
#> #   recoveries <int>, key_passes <int>, tackles <int>, winning_goals <int>,
#> #   attempted_passes <int>, completed_passes <int>, penalties_conceded <int>,
#> #   big_chances_missed <int>, errors_leading_to_goal <int>,
#> #   errors_leading_to_goal_attempt <int>, tackled <int>, offside <int>,
#> #   target_missed <int>, fouls <int>, dribbles <int>