Detailed data for a player in the current FPL season

Returns a tibble containing gameweek-level data for a given player in the current FPL season (season has to be in-play).

playerDetailed(player_id)

Arguments

player_id
id field from players tibble for a desired player.

Examples

playerDetailed(player_id = 1)

#> # A tibble: 1 x 54
#>      id         kickoff_time kickoff_time_formatted team_h_score team_a_score
#>   <int>                <chr>                  <chr>        <int>        <int>
#> 1     1 2017-08-11T18:45:00Z           11 Aug 19:45            4            3
#> # ... with 49 more variables: was_home <lgl>, round <int>, total_points <int>,
#> #   transfers_balance <int>, selected <int>, transfers_in <int>,
#> #   transfers_out <int>, loaned_in <int>, loaned_out <int>, minutes <int>,
#> #   goals_scored <int>, assists <int>, clean_sheets <int>,
#> #   goals_conceded <int>, own_goals <int>, penalties_saved <int>,
#> #   penalties_missed <int>, yellow_cards <int>, red_cards <int>, saves <int>,
#> #   bonus <int>, bps <int>, influence <dbl>, creativity <dbl>, threat <dbl>,
#> #   ict_index <dbl>, ea_index <int>, open_play_crosses <int>,
#> #   big_chances_created <int>, clearances_blocks_interceptions <int>,
#> #   recoveries <int>, key_passes <int>, tackles <int>, winning_goals <int>,
#> #   attempted_passes <int>, completed_passes <int>, penalties_conceded <int>,
#> #   big_chances_missed <int>, errors_leading_to_goal <int>,
#> #   errors_leading_to_goal_attempt <int>, tackled <int>, offside <int>,
#> #   target_missed <int>, fouls <int>, dribbles <int>, fixture <int>,
#> #   opponent_team <chr>, price <dbl>, player_id <int>

playerDetailed(player_id = 54)

#> # A tibble: 1 x 54
#>      id         kickoff_time kickoff_time_formatted team_h_score team_a_score
#>   <int>                <chr>                  <chr>        <int>        <int>
#> 1    54 2017-08-12T16:30:00Z           12 Aug 17:30            0            2
#> # ... with 49 more variables: was_home <lgl>, round <int>, total_points <int>,
#> #   transfers_balance <int>, selected <int>, transfers_in <int>,
#> #   transfers_out <int>, loaned_in <int>, loaned_out <int>, minutes <int>,
#> #   goals_scored <int>, assists <int>, clean_sheets <int>,
#> #   goals_conceded <int>, own_goals <int>, penalties_saved <int>,
#> #   penalties_missed <int>, yellow_cards <int>, red_cards <int>, saves <int>,
#> #   bonus <int>, bps <int>, influence <dbl>, creativity <dbl>, threat <dbl>,
#> #   ict_index <dbl>, ea_index <int>, open_play_crosses <int>,
#> #   big_chances_created <int>, clearances_blocks_interceptions <int>,
#> #   recoveries <int>, key_passes <int>, tackles <int>, winning_goals <int>,
#> #   attempted_passes <int>, completed_passes <int>, penalties_conceded <int>,
#> #   big_chances_missed <int>, errors_leading_to_goal <int>,
#> #   errors_leading_to_goal_attempt <int>, tackled <int>, offside <int>,
#> #   target_missed <int>, fouls <int>, dribbles <int>, fixture <int>,
#> #   opponent_team <chr>, price <dbl>, player_id <int>